backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
langchain-groq
python-jose